requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=1.4.25
python-jose>=3.3.0
pydantic[email]
passlib>=1.7.4
python-multipart
bcrypt>=3.2.0